+++ krcg_cli/subcommands/seating.py
@@ -1,17 +1,63 @@
+import argparse
 import collections
 import functools
 import itertools
 import multiprocessing
+import random
 import sys
 
 from krcg import seating
 
 
 def add_parser(parser):
-    parser = parser.add_parser("seating", help="compute optimal seating")
+    parser = parser.add_parser(
+        "seating",
+        help="compute optimal seating",
+        formatter_class=argparse.RawDescriptionHelpFormatter,
+        description=(
+            """Optimal seating is useful for tournament play.
+An optimal seating follows the rules established by the Rules Director, see
+https://groups.google.com/g/rec.games.trading-cards.jyhad/c/4YivYLDVYQc/m/CCH-ZBU5UiUJ
+
+The output is a normalised comma-separated list of players, one line per round
+Tables of 5 are first, tables of 4, if any, are last.
+For example, 1,2,3,4,5,6,7,8,9,10,11,12,13 unambiguously means:
+[1, 2, 3, 4, 5], [6, 7, 8, 9], [10, 11, 12, 13]
+
+Use the -v option to display the table structure for each round.
+The comma-separated normal form can be used as input of the command
+to indicate rounds that have already been played.
+This allows for players list modifications during a tournament. For example:
+
+$ krcg seating --played 1,2,3,4,5,6,7,8,9 --remove 4
+
+will output a new seating with the played round left untouched,
+and the next 2 rounds without player 4.
+
+Seating for 6, 7 or 11 players required multiple intertwined rounds.
+For example, for 6 players to play 2 rounds each, 3 rounds are required
+with some players sitting out on each of them:
+
+$ krcg seating --rounds 2 6
+2,6,1,4
+4,1,5,3
+3,5,6,2
+
+This can also be the case when you remove players and come down to 6, 7, or 11 players.
+The command will accomodate if they are enough rounds left for intertwined rounds:
+
+$ krcg seating --played 1,2,3,4,5,6,7,8,9 --remove 4 5
+
+Note that when you began to play such intertwined rounds, you cannot modify
+the players list in the middle of them. Trying to use the command in this case will
+yield unusable seatings listing only the players of the last round.
+"""
+        ),
+    )
     parser.add_argument(
         "players",
         type=int,
+        nargs="?",
         metavar="PLAYERS",
         help=("Number of players."),
     )
@@ -20,10 +66,7 @@ def add_parser(parser):
         "--rounds",
         type=int,
         default=3,
-        help=(
-            "Number of rounds, "
-            "if this argument is used the [players] argument can only appear once"
-        ),
+        help=("Number of rounds"),
     )
     parser.add_argument(
         "-i",
@@ -33,21 +76,32 @@ def add_parser(parser):
         help="Number of iterations to use (less is faster but may yield worse results)",
     )
     parser.add_argument(
-        "--remove",
-        type=lambda s: tuple((int(x) for x in s.split(":"))),
+        "-v",
+        "--verbose",
+        action="store_true",
+        help="Display seating tables and statistics",
+    )
+    parser.add_argument(
+        "-p",
+        "--played",
+        type=lambda s: [int(x) for x in s.split(",") if x],
         nargs="*",
-        metavar="ROUND:PLAYER",
-        help=(
-            "Remove given player, starting from given round. "
-            "Format is '[round]:[player]'"
-        ),
+        metavar="ROUND",
+        help="Rounds that have already been played",
+    )
+    parser.add_argument(
+        "--remove",
+        type=int,
+        nargs="*",
+        metavar="PLAYER",
+        help="Remove given players",
     )
     parser.add_argument(
         "--add",
         type=int,
         nargs="*",
-        metavar="ROUND",
-        help="Add a player, starting from given round.",
+        metavar="PLAYER",
+        help="Add given players",
     )
     parser.set_defaults(func=seat)
 
@@ -66,98 +120,89 @@ GROUPS = {
 }
 
 
-def progression(iterations, step):
+def progression(iterations, step, **kwargs):
     print(f"\t{step / iterations * 100:.0f}%", file=sys.stderr, end="\r")
 
 
-def seat(args):
-    players = args.players
-    if args.rounds:
-        permutations = seating.permutations(players, args.rounds)
+def seat(options):
+    if options.players and options.played:
+        print(
+            "the [played] and [players] arguments cannot be used both", file=sys.stderr
+        )
+        return 1
+    if not options.players and not options.played:
+        print("one of [played] or [players] arguments must be used", file=sys.stderr)
+        return 1
+
+    if options.players:
+        players = set(range(1, options.players + 1))
+        rounds = options.rounds
     else:
-        permutations = [list(range(1, p + 1)) for p in players]
+        players = set(options.played[-1])
+        rounds = options.rounds - len(options.played)
 
-    # addition/removal is a bit tricy
-    # we add/remove all players at once: 10<->11 is difficult, but 10<->12 works
-    # we use seating.permutations: seating 6, 7 or 11 players
-    # might still be doable by adding more rounds, if there are enough rounds left.
-    # if you began interlaced rounds for 6, 7 or 11 players,
-    # additions and removals are not doable.
-    if args.players in [6, 7, 11] and args.add:
-        print(f"cannot add players to {args.players} players", file=sys.stderr)
-        return 1
-    additions = collections.Counter(args.add or [])
-    for round_, count in sorted(additions.items()):
-        players += count
-        permutations = permutations[: round_ - 1] + seating.permutations(
-            players, args.rounds - round_ + 1
-        )
-    if args.players in [6, 7, 11] and args.remove:
-        print(f"cannot remove players from {args.players} players", file=sys.stderr)
-        return 1
-
-    # removal
-    removals = collections.defaultdict(set)
-    for round_, player in args.remove or []:
-        removals[round_].add(player)
-    removed_players = []
-    for round_, removed in sorted(removals.items()):
-        # make sure we are removing players that were playing
-        absent = removed - set(
-            p for p in itertools.chain.from_iterable(permutations[round_ - 1 :])
-        )
-        if absent:
+    for player in options.add or []:
+        if player in players:
             print(
-                f"trying to remove {absent} starting from round {round_}, "
-                "but they're absent",
+                f"trying to add {player} but they are already in",
+                file=sys.stderr,
+            )
+        players.add(player)
+    for player in options.remove or []:
+        if player not in players:
+            print(
+                f"trying to remove {player} but they absent",
                 file=sys.stderr,
             )
             return 1
-        players -= len(removed)
-        removed_players.extend(removed)
-        try:
-            permutations = permutations[: round_ - 1] + seating.permutations(
-                players, args.rounds - round_ + 1
-            )
-        except RuntimeError:
-            print(
-                f"removal in round {round_} " "yields an invalid number of players.",
-                file=sys.stderr,
-            ),
-            return 1
-        # since seating.permutations assigns a "normal" range of player numbers
-        # we need to replace the removed numbers by the right ones
-        for i, player in enumerate(removed_players):
-            for permutation in permutations[round_ - 1 :]:
-                try:
-                    # careful, this has to be stable: removed_players list is in round order
-                    permutation[permutation.index(player)] = (
-                        players + len(removed_players) - i
-                    )
-                except ValueError:
-                    pass
+        players.remove(player)
 
     try:
-        cpus = multiprocessing.cpu_count()
-    except NotImplementedError:
-        cpus = 1
-    with multiprocessing.Pool(processes=cpus) as pool:
-        results = [
-            pool.apply_async(
-                seating.optimise,
-                (
-                    permutations,
-                    args.iterations,
-                    functools.partial(progression, args.iterations),
-                ),
-            )
-            for _ in range(cpus)
-        ]
-        rounds, score = min((r.get() for r in results), key=lambda x: x[1].total)
+        permutations = seating.permutations(len(players), rounds)
+    except RuntimeError:
+        print(
+            "seating cannot be arranged - more rounds or players required",
+            file=sys.stderr,
+        )
+        return 1
 
+    players = list(players)
+    random.shuffle(players)
+    players = {i: p for i, p in enumerate(players, 1)}
+    permutations = (options.played or []) + [
+        [players[i] for i in permutation] for permutation in permutations
+    ]
+
+    if rounds > 0:
+        try:
+            cpus = multiprocessing.cpu_count()
+        except NotImplementedError:
+            cpus = 1
+        with multiprocessing.Pool(processes=cpus) as pool:
+            results = [
+                pool.apply_async(
+                    seating.optimise,
+                    kwds=dict(
+                        permutations=permutations,
+                        iterations=options.iterations,
+                        callback=functools.partial(progression, options.iterations),
+                        fixed=max(1, len(options.played or [])),
+                        ignore=set((options.add or []) + (options.remove or [])),
+                    ),
+                )
+                for _ in range(cpus)
+            ]
+            rounds, score = min((r.get() for r in results), key=lambda x: x[1].total)
+    else:
+        rounds = [seating.Round(p) for p in permutations]
+        score = seating.score_rounds(rounds)
+    for round_ in rounds:
+        print(",".join(str(p) for p in itertools.chain.from_iterable(round_)))
+    if not options.verbose:
+        return 0
+    print("--------------------------------- details ---------------------------------")
     for i, round_ in enumerate(rounds, 1):
         print(f"Round {i}: {round_}")
-    print(f"Total score: {score.total:.2f}")
     for index, (code, label, _) in enumerate(seating.RULES):
         s = f"{code} {score.rules[index]:6.2f} "
         if score.rules[index]: